crm-ai-service/requirements.txt
fastapi>=0.109.0
uvicorn>=0.27.0
python-multipart>=0.0.9
pdfminer.six>=20231228
python-docx>=1.1.0
pytest>=8.0.0
